star_navi_social/pyproject.toml
[tool.poetry]
name = "starnavi-social"
version = "0.1.0"
description = ""
authors = ["Klukvas <pavlenkoandrey56@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
pydantic = "^2.5.3"
sqlalchemy = "^2.0.25"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
python-dotenv = "^1.0.1"
email-validator = "^2.1.0.post1"
greenlet = "^3.0.3"
pydantic-settings = "^2.1.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
pyjwt = "^2.8.0"
bcrypt = "^4.1.2"
python-dateutil = "^2.8.2"
psycopg2-binary = "^2.9.9"
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.1.0"
flake8 = "^7.0.0"
pytest = "^8.0.0"
httpx = "^0.26.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q"
markers = ["registration"]


[tool.black]
line-length = 120
target-version = ['py311']